requirements.txt
argcomplete==1.10.0
astroid==2.3.2
beautifulsoup4==4.8.0
bert-serving-client==1.9.8
cachetools==3.1.1
certifi==2019.9.11
cffi==1.13.1
chardet==3.0.4
Click==7.0
cryptography==2.8
decorator==4.4.1
docusign-esign==3.0.0
docx2txt==0.8
EbookLib==0.17.1
extract-msg==0.23.1
Flask==1.1.1
Flask-Sockets==0.2.1
gevent==1.4.0
gevent-websocket==0.10.1
google-resumable-media==0.4.1
googleapis-common-protos==1.6.0
greenlet==0.4.15
grpcio==1.24.3
idna==2.8
IMAPClient==2.1.0
img2pdf==0.3.3
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.3
joblib==0.14.0
lazy-object-proxy==1.4.2
lxml==4.4.1
MarkupSafe==1.1.1
mccabe==0.6.1
networkx==2.4
nltk==3.4.5
nose==1.3.7
numpy==1.17.3
ocrmypdf==9.0.3
olefile==0.46
pdf2text==1.0.0
pdfminer==20191020
pdfminer.six==20181108
pendulum==2.0.5
pikepdf==1.6.5
Pillow==6.2.1
protobuf==3.10.0
pyasn1==0.4.7
pyasn1-modules==0.2.7
pycparser==2.19
pycryptodome==3.9.0
PyJWT==1.7.1
pylint==2.4.3
pytesseract==0.3.0
python-dateutil==2.8.0
python-pptx==0.6.18
pytz==2019.3
pytzdata==2019.3
pyzmq==18.1.0
reportlab==3.5.31
requests==2.22.0
rsa==4.0
scikit-learn==0.21.3
scipy==1.3.1
six==1.12.0
sklearn==0.0
sortedcontainers==2.1.0
soupsieve==1.9.4
SpeechRecognition==3.8.1
textract==1.6.3
tqdm==4.36.1
typed-ast==1.4.0
tzlocal==1.5.1
urllib3==1.25.6
Wand==0.5.7
Werkzeug==0.16.0
wrapt==1.11.2
xlrd==1.2.0
XlsxWriter==1.2.2
